refactor(gateway): split task supervisors
